switched model base to illuminate eloquent ORM
