+++ controllers/base.php
@@ -1,6 +1,6 @@
 <?php
 namespace Controller;
-class Base{
+abstract class Base{
 	protected $app;
 	public function __construct(){
 		$this->app = \Slim\Slim::getInstance();
