+++ index.php
@@ -1,8 +1,8 @@
 <?php
-require 'vendor/autoload.php';
+//require 'vendor/autoload.php';
 require 'config.php';
 
-
+/*
 // Set the view parser extensions.
 $view = new \Slim\Views\Twig();
 $view->parserExtensions = array(new \Slim\Views\TwigExtension());
@@ -11,13 +11,13 @@ $app = new \Slim\Slim(array('mode' => 'development',
 							'debug' => true,
 							'view'=> $view,
 							'templates.path'=> 'views'
-							));
+							));*/
 //$app->add(new \Middleware\AuthMiddleware());
 
 //dev branch test
 $app->get('/', 'Controller\Home:display');
 $app->post('/home','Controller\Home:login')->name('UserLogin');
-//$app->get('/home',  'Controller\Home:loginForm')->name('UserLoginForm');
+$app->get('/home',  'Controller\Home:loginForm')->name('UserLoginForm');
 $app->post('/signup','Controller\Home:signUp')->name('UserSignup');
 $app->get('/edit', 'Controller\User:edit')->name('editAccount');
 //$app->post('/home', 'Controller\Home:test')->name('HomeTest');
